chore: restructure dropdown-menu
