+++ src/dropdown-menu/DropdownMenuDivider.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function DropdownDivider(props) {
+  return <div className="pe-dropdown-menu__divider" role="seperator"></div>
+};
+
+export default DropdownMenuDivider;
