+++ src/dropdown-menu/DropdownMenuItem.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function DropdownItem(props) {
+  return <div {...props} className="pe-dropdown-menu__menu-item">{props.children}</div>
+};
+
+export default DropdownMenuItem;
