+++ src/index.js
@@ -5,4 +5,6 @@ export { default as CardContent } from './card/CardContent';
 export { default as Label } from './Label';
 export { default as Lead } from './Lead';
 export { default as Title } from './Title';
-export { default as Dropdown, DropdownDivider, DropdownItem } from './Dropdown';
+export { default as DropdownMenu } from './dropdown-menu/DropdownMenu';
+export { default as DropdownMenuItem } from './dropdown-menu/DropdownMenuItem';
+export { default as DropdownMenuDivider } from './dropdown-menu/DropdownMenuDivider';
